Add mkiso script to create iso bundled with install script
I find it quite annoying typing into the terminal the full path to the
install script on github and less annoying to type a series of fiddly
commands in instead.

The install script is copied to the root of the usb drive which means
that on the install candidate we need to mount the usb drive and copy
the file out which is a bit messy. I'd like it if I could have put the
install script directly inside the airootfs/squashfs
image (/arch/x86_64/airootfs.sfs) that's mounted at / instead but the
mac implementation of the squashfs tools don't seem to be playing
ball.