Allow remote ansible playbook triggering
This adds hosts config so that I can trigger ansible remotely, it is
hardcoded with references to my ssh private key on my development
machine and a local private 10.x.x.x ip address for the server that is
also only available locally. The ip address isn't static so I have to
hope it doesn't change too often. The local hosts file has been
renamed hosts_xps_local.ini so that the install script can still
bootstrap the server.

To allow remote ansible execution I've added my daniel user to the
sudoers explicitly with NOPASSWD set. I had previously added the user
to the wheel group for sudo access and this is left in place to
provide redundancy.