Separate chroot and post chroot setup
I've had some difficulty writing ansible config that runs both in
chroot environment and then in a normal environment, so here's a
commit to split them up.

A new playbook xps_chroot.yml is added that sets up the minimum
environment required for the next stage of the setup, this minimum
environment sets up the network, sshd, and my user.

Then the install process has a second section that can be run
remotely, this sets up everything else and is defined in the original
xps.yml playbook.