Add Plex Media Server
This change adds Plex Media Server and sets it up to share
/mnt/storage/media, the create directory tasks for this and config are
commented out while I test a bit more to ensure I don't accidentally
wipe everything with some plex commands.

Docker is added too here, it's linked as a dependency from the Plex
role but I've also included in explicitly in the xps playbook because
I have plans for it as an integral component.