Add makemkv
I need to double check that this installs correctly, the manual
process to install makemkv pauses on a license screen that needs user
input to continue. Such it is that I haven't tested the automated
install yet and I'm afraid that it will also block and prevent the
install from completing and perhaps also prevent the next parts of the
system from being configured too.