Configure external disks
This adds config for my usb seagate external drive and my Terra-Master
DAS with a RAID1 array with a single disk (second is on its way
later).

The seagate drive is mount at /mnt/seagate and the DAS / RAID array is
mounted at /mnt/tmdas.

The Plex data directory mount has been changed to /mnt/tmdas/video
which is the new location now I've been able to rip more discs and
have the requirement for that extra space.